datastructures: allow smaller atomic counters
This change allows for the use of smaller atomic counters. Rust 1.34 stabilized
fixed width atomics for u8, u16, u32, and u64 which may be more suitable for
specific use-cases than atomic usize. Particularly, for histograms covering wide
ranges of values, but where any given bucket is unlikely to have a large count,
reducing from atomic usize (effectively 64bit atomics on x86_64) to u32 could
yield appreciable savings in memory footprint.

To support this change, we make `Counter` generic across u8, u16, u32, and u64.
Histograms and heatmaps then are able to use these smaller uints to hold bucket
counters. This change also updates dependent crates within the project to make
use of the changes to the datastructures.

+++ rpc-perf/src/codec/redis.rs
@@ -53,7 +53,7 @@ impl Codec for Redis {
                 let key = command.key().unwrap();
                 if let Some(recorder) = self.common.recorder() {
                     recorder.increment("commands/get");
-                    recorder.distribution("keys/size", key.len());
+                    recorder.distribution("keys/size", key.len() as u64);
                 }
                 self.codec.get(buf, key);
             }
@@ -62,8 +62,8 @@ impl Codec for Redis {
                 let value = command.value().unwrap();
                 if let Some(recorder) = self.common.recorder() {
                     recorder.increment("commands/set");
-                    recorder.distribution("keys/size", key.len());
-                    recorder.distribution("values/size", value.len());
+                    recorder.distribution("keys/size", key.len() as u64);
+                    recorder.distribution("values/size", value.len() as u64);
                 }
                 self.codec.set(buf, key, value, command.ttl());
             }
